Move redact_uri
