+++ src/layman/layer/db/table.py
@@ -1,12 +1,13 @@
 import logging
 from psycopg2 import sql
 
+import requests_util.url_util
 from db import util as db_util
 from layman import settings, patch_mode, util as layman_util
 from layman.common import empty_method, empty_method_returns_none, empty_method_returns_dict
 from layman.http import LaymanError
 from . import get_internal_table_name
-from .. import LAYER_TYPE, util as layer_util
+from .. import LAYER_TYPE
 
 logger = logging.getLogger(__name__)
 PATCH_MODE = patch_mode.DELETE_IF_DEPENDANT
@@ -35,7 +36,7 @@ def get_layer_info(workspace, layername,):
                     'schema': table_uri.schema,
                     'table': table_uri.table,
                     'geo_column': table_uri.geo_column,
-                    'external_uri': layer_util.redact_uri(table_uri.db_uri_str),
+                    'external_uri': requests_util.url_util.redact_uri(table_uri.db_uri_str),
                     'status': 'NOT_AVAILABLE',
                     'error': 'Cannot connect to DB.',
                 }
@@ -56,13 +57,13 @@ def get_layer_info(workspace, layername,):
                 'geo_column': table_uri.geo_column,
             }
             if layer_info['original_data_source'] == settings.EnumOriginalDataSource.TABLE.value:
-                result['db']['external_uri'] = layer_util.redact_uri(table_uri.db_uri_str)
+                result['db']['external_uri'] = requests_util.url_util.redact_uri(table_uri.db_uri_str)
         elif layer_info['original_data_source'] == settings.EnumOriginalDataSource.TABLE.value:
             result['db'] = {
                 'schema': table_uri.schema,
                 'table': table_uri.table,
                 'geo_column': table_uri.geo_column,
-                'external_uri': layer_util.redact_uri(table_uri.db_uri_str),
+                'external_uri': requests_util.url_util.redact_uri(table_uri.db_uri_str),
                 'status': 'NOT_AVAILABLE',
                 'error': 'Table does not exist.',
             }
