+++ src/layman/layer/util.py
@@ -443,17 +443,6 @@ WHERE
     return result
 
 
-def redact_uri(uri):
-    url_components = parse.urlparse(uri)
-    if url_components.username or url_components.password:
-        hostname = f'{url_components.hostname}:{url_components.port}' if url_components.port else url_components.hostname
-        url_components = url_components._replace(
-            netloc=f"{url_components.username}@{hostname}",
-        )
-
-    return url_components.geturl()
-
-
 def set_wfs_wms_status_after_fail(workspace, name):
     keys = ['wfs', 'wms', 'style', 'geodata_type']
     publ_info = layman_util.get_publication_info(workspace, LAYER_TYPE, name, context={'keys': keys})
